Revert "add rolling update daemonset existing pod adoption integration test"
